### 2023-10-17 #614 @osorensen
galamm: Generalized Additive Latent and Mixed Models
Submitting Author Name: Øystein Sørensen
Submitting Author Github Handle: @osorensen
Repository:  https://github.com/LCBC-UiO/galamm
Submission type: Pre-submission

---

```
Package: galamm
Title: Generalized Additive Latent and Mixed Models
Version: 0.1.1.9000
    person(given = "Øystein",
           family = "Sørensen",
           email = "oystein.sorensen@psykologi.uio.no",
    person(given = "Douglas", family = "Bates", role = "ctb"),
    person(given = "Ben", family = "Bolker", role = "ctb"),
    person(given = "Martin", family = "Maechler", role = "ctb"),
    person(given = "Allan", family = "Leal", role = "ctb"),
    person(given = "Fabian", family = "Scheipl", role = "ctb"),
    person(given = "Steven", family = "Walker", role = "ctb"),
    person(given = "Simon", family = "Wood", role = "ctb")
           )
Description: Estimates generalized additive latent and
    mixed models using maximum marginal likelihood,
    as defined in Sorensen et al. (2023)
    <doi:10.1007/s11336-023-09910-z>, which is an extension of Rabe-Hesketh and
    Skrondal (2004)'s unifying framework for multilevel latent variable
    modeling <doi:10.1007/BF02295939>. Efficient computation is done using sparse
    matrix methods, Laplace approximation, and automatic differentiation. The
    framework includes generalized multilevel models with heteroscedastic
    residuals, mixed response types, factor loadings, smoothing splines,
    crossed random effects, and combinations thereof. Syntax for model
    formulation is close to 'lme4' (Bates et al. (2015)
    <doi:10.18637/jss.v067.i01>) and 'PLmixed' (Rockwood and Jeon (2019)
    <doi:10.1080/00273171.2018.1516541>).
    lme4,
    Matrix,
    memoise,
    mgcv,
    nlme,
    Rcpp,
    Rdpack,
    stats
    Rcpp,
    RcppEigen
    covr,
    gamm4,
    PLmixed,
VignetteBuilder: knitr, rmarkdown
RdMacros: Rdpack
NeedsCompilation: yes
SystemRequirements: C++17
```

The package estimates mixed effect and latent variable models which may include smooth nonlinear terms. Pretty obvious that it's regression software.
No. `pkgcheck::pkgcheck()` says the package is ready to be submitted, but I have the same issue with the `autotest` package as described [here](https://github.com/ropensci-review-tools/autotest/issues/83). Running `x <- autotest_package(test = TRUE)` gives 6 errors, 5 warnings, and 21 diagnostic messages, as shown below:

```r
> as.data.frame(x) %>% group_by(type) %>% count()
# A tibble: 3 × 2
# Groups:   type [3]
  type           n
  <chr>      <int>
1 diagnostic    21
2 error          6
3 warning        5
```
For example, the first error has the following entry in the `content` column, and the other errors are similar.
```r
> as.data.frame(x)[1, "content"]
[1] ":quote(y ~ x + (1 | id)), weights = base::quote(~(1 | : family_mapping must be a vector."
```
I don't know what this means, and I'm not able to figure it out by reading the documentation and vignettes of the `autotest` package. Anyhow, since the **srr** package is supposed to be used after **autotest** returns `NULL`, I haven't been able to try it.
The target audience is applied statisticians and quantitative scientists, particularly those working on the social sciences. The package is motivated by longitudinal studies in cognitive neuroscience, but it is applicable wherever a measurement model (of factor analysis type) needs to be combined with hierarchical modeling.
None that I am aware of. It is possible to implement these types of models using, e.g., Stan or JAGS, and potentially some using OpenMx, but that requires a lot more programming. The package does have overlapping features with other R packages, including lme4, mgcv, PLmixed, gamlss, and lavaan, but the aim of this package is to provide functionality not supported by those packages.

I have written this package myself, but I have imported some limited amount of code from packages `lme4` and `mgcv`, as well as the `C++` library autodiff. In accordance with the policy stated in [Chapter 1.9 of the rOpenSci online book](https://devguide.ropensci.org/building.html?q=contributor#authorship), I have added the authors of those packages as contributors in `DESCRIPTION`.
