### 2023-10-17 ee66773
recompiled scaling vignette, got rid of ggplot2 in Suggests
