+++ vignettes/semiparametric.Rmd
@@ -18,8 +18,6 @@ vignette: >
 ```r
 library(galamm)
 library(gamm4)
-library(ggplot2)
-theme_set(theme_bw())
 ```
 
 This vignette describes how to use `galamm` to estimate latent variable models with smooth terms, or equivalently, generalized additive mixed models with factor structures. The examples are based on Section 4 and 5 in @sorensenLongitudinalModelingAgeDependent2023, but as we cannot share the data, we have instead simulated somewhat simpler datasets that will be used. We will gradually add complexity, starting with a simple generalized additive mixed model. Please refer to the [introductory vignette](https://lcbc-uio.github.io/galamm/articles/galamm.html) for an overview of the statistical models.
@@ -55,9 +53,12 @@ Each subject in this dataset has been measured eight times, and we can plot the 
 
 
 ```r
-ggplot(dat, aes(x = x, y = y, group = id)) +
-  geom_point(size = .1) +
-  geom_line(alpha = .3)
+plot(dat$x, dat$y, type = "n", xlab = "x", ylab = "y")
+for(i in unique(dat$id)) {
+  dd <- dat[dat$id == i, ]
+  lines(dd$x, dd$y, col = "gray")
+}
+points(dat$x, dat$y, pch = 20, lwd = .05)
 ```
 
 ![Plot of data for domain 1 and item 11.](semiparametric-spaghetti-plot-1.png)
@@ -627,11 +628,9 @@ The summary shows that we have recovered the true values of the factor loadings 
 ```r
 summary(mod_byvar)
 #> GALAMM fit by maximum marginal likelihood.
-#> Formula: y ~ domain + sl(x, by = domain, factor = c("ability1", "ability3")) +  
-#>     (0 + domain1:ability1 + domain3:ability3 | id/timepoint)
+#> Formula: y ~ domain + sl(x, by = domain, factor = c("ability1", "ability3")) +      (0 + domain1:ability1 + domain3:ability3 | id/timepoint)
 #>    Data: dat
-#> Control: galamm_control(optim_control = list(factr = 1e+09, trace = 3,  
-#>     REPORT = 50, maxit = 1000))
+#> Control: galamm_control(optim_control = list(factr = 1e+09, trace = 3,      REPORT = 50, maxit = 1000))
 #> 
 #>      AIC      BIC   logLik deviance df.resid 
 #>  -2398.7  -2266.9   1217.3  -2434.7    11182 
@@ -771,11 +770,9 @@ We can look at the model summary:
 ```r
 summary(mod_byvar_mixed)
 #> GALAMM fit by maximum marginal likelihood.
-#> Formula: y ~ domain + sl(x, by = domain, factor = c("ability1", "ability2")) +  
-#>     (0 + domain1:ability1 + domain2:ability2 | id)
+#> Formula: y ~ domain + sl(x, by = domain, factor = c("ability1", "ability2")) +      (0 + domain1:ability1 + domain2:ability2 | id)
 #>    Data: dat
-#> Control: galamm_control(optim_control = list(factr = 1e+09, trace = 3,  
-#>     REPORT = 30, maxit = 1000))
+#> Control: galamm_control(optim_control = list(factr = 1e+09, trace = 3,      REPORT = 30, maxit = 1000))
 #> 
 #>      AIC      BIC   logLik deviance df.resid 
 #>   9476.1   9566.9  -4725.0  98981.3     7987 
