### 2023-10-17 cf1be84
radical simplification of lambda_mapping_X
