+++ R/misc.R
@@ -26,7 +26,7 @@
 #' # NA's in the original formulation have been replaced by 1's:
 #' res$lambda
 setup_factor <- function(load.var, lambda, factor, data) {
-  if (!is.null(load.var) && (!is.character(load.var) || length(load.var) != 1)) {
+  if (!is.null(load.var) && (!is.character(load.var))) {
     stop("load.var must be a character of length one.")
   }
 
