### 2023-10-18 d13a0e6
returned fast when there are no factor mappings
