### 2023-10-18 c3983bd
fixed error in formula.galammm
