+++ pkgdown.yml
@@ -11,7 +11,7 @@ articles:
   optimization: optimization.html
   scaling: scaling.html
   semiparametric: semiparametric.html
-last_built: 2023-10-17T12:44Z
+last_built: 2023-10-18T07:09Z
 urls:
   reference: https://lcbc-uio.github.io/galamm/reference
   article: https://lcbc-uio.github.io/galamm/articles
