### 2023-10-18 53d3376
seem to have fixed a convergence issue
