+++ tests/testthat/test-galamm-setup.R
@@ -8,9 +8,10 @@ test_that("wrong input is handled properly", {
       data = dat,
       family = binomial,
       load.var = "item",
-      lambda = list(matrix(c(1, NA, NA), ncol = 1)),
-      factor = list("loading")
-    )
+      lambda = matrix(c(1, NA, NA), ncol = 1),
+      factor = "loading"
+    ),
+    "lambda matrix must contain one row for each element in load.var"
   )
 
   expect_error(
@@ -20,8 +21,8 @@ test_that("wrong input is handled properly", {
       data = dat,
       family = binomial,
       load.var = "item",
-      lambda = list(matrix(c(1, NA), ncol = 1)),
-      factor = list("loading"),
+      lambda = matrix(c(1, NA), ncol = 1),
+      factor = "loading",
       start = list(phi = 1)
     ),
     "Unknown names in initial value list"
@@ -34,8 +35,8 @@ test_that("wrong input is handled properly", {
       data = dat,
       family = binomial,
       load.var = "item",
-      lambda = list(matrix(c(1, NA), ncol = 1)),
-      factor = list("loading"),
+      lambda = matrix(c(1, NA), ncol = 1),
+      factor = "loading",
       start = list(theta = 1)
     ),
     "Wrong number of elements"
@@ -48,8 +49,8 @@ test_that("wrong input is handled properly", {
       data = dat,
       family = binomial,
       load.var = "item",
-      lambda = list(matrix(c(1, NA), ncol = 1)),
-      factor = list("loading"),
+      lambda = matrix(c(1, NA), ncol = 1),
+      factor = "loading",
       start = list(beta = rep(1, 10))
     ),
     "Wrong number of elements"
@@ -62,8 +63,8 @@ test_that("wrong input is handled properly", {
       data = dat,
       family = binomial,
       load.var = "item",
-      lambda = list(matrix(c(1, NA), ncol = 1)),
-      factor = list("loading"),
+      lambda = matrix(c(1, NA), ncol = 1),
+      factor = "loading",
       start = list(lambda = rep(1, 5))
     ),
     "Wrong number of elements"
@@ -76,8 +77,8 @@ test_that("wrong input is handled properly", {
       data = dat,
       family = binomial,
       load.var = "item",
-      lambda = list(matrix(c(1, NA), ncol = 1)),
-      factor = list("loading"),
+      lambda = matrix(c(1, NA), ncol = 1),
+      factor = "loading",
       start = list(weights = rep(1, 5))
     ),
     "Wrong number of elements"
@@ -92,7 +93,6 @@ test_that("wrong input is handled properly", {
     "Multiple grouping terms in weights not yet implemented."
   )
 
-
   newdat <- dat
   newdat$loading <- 1
   expect_error(
@@ -102,14 +102,15 @@ test_that("wrong input is handled properly", {
       data = newdat,
       family = binomial,
       load.var = "item",
-      lambda = list(matrix(c(1, NA, NA), ncol = 1)),
-      factor = list("loading")
+      lambda = matrix(c(1, NA, NA), ncol = 1),
+      factor = "loading"
     ),
     "Factor already a column in data"
   )
 
   expect_error(
-    mod <- galamm(formula = y ~ (1 | id), data = subset(dat, FALSE))
+    mod <- galamm(formula = y ~ (1 | id), data = subset(dat, FALSE)),
+    "No data, nothing to do."
   )
 
   expect_error(
@@ -218,7 +219,7 @@ test_that("wrong input is handled properly", {
         load.var = 1
       )
     },
-    "load.var must be a character of length one"
+    "load.var must be NULL or a character of length one"
   )
 
[…]
