+++ R/factor_loadings.R
@@ -53,7 +53,7 @@ factor_loadings.galamm <- function(object) {
     return(invisible(NULL))
   }
 
-  lambda_tmp_est <- lambda_tmp_se <- object$parameters$lambda_dummy[[1]]
+  lambda_tmp_est <- lambda_tmp_se <- object$parameters$lambda_dummy
   lambda_tmp_se[lambda_tmp_se %in% c(0, 1)] <- NA_real_
 
   lambda_tmp_est[lambda_tmp_est > 1] <-
@@ -61,7 +61,7 @@ factor_loadings.galamm <- function(object) {
   lambda_tmp_se[!is.na(lambda_tmp_se)] <-
     sqrt(diag(vcov(object, parm = "lambda")))
 
-  nn <- nrow(object$parameters$lambda_dummy[[1]])
+  nn <- nrow(object$parameters$lambda_dummy)
   ret <- matrix(rbind(lambda_tmp_est, lambda_tmp_se),
     nrow = nrow(lambda_tmp_est),
     dimnames = list(
@@ -72,7 +72,7 @@ factor_loadings.galamm <- function(object) {
 
   lix <- length(object$parameters$lambda_interaction_inds)
   if (lix > 0) {
-    vars <- unlist(lapply(object$model$factor_interactions[[1]], function(x) {
+    vars <- unlist(lapply(object$model$factor_interactions, function(x) {
       attr(stats::terms(x), "term.labels")
     }))
 
