+++ R/factor_loadings.R
@@ -42,7 +42,7 @@ factor_loadings <- function(object) {
 #' # Estimate model
 #' mod <- galamm(y ~ item + (0 + ability | sid) + (0 + ability | school),
 #'   data = IRTsub, family = binomial, load.var = "item",
-#'   factor = list("ability"), lambda = list(loading_matrix)
+#'   factor = "ability", lambda = loading_matrix
 #' )
 #'
 #' # Show estimated factor loadings, with standard errors
