+++ R/formula.R
@@ -26,9 +26,9 @@
 #'   data = mresp,
 #'   family = families,
 #'   family_mapping = family_mapping,
-#'   factor = list("level"),
+#'   factor = "level",
 #'   load.var = "itemgroup",
-#'   lambda = list(loading_matrix)
+#'   lambda = loading_matrix
 #' )
 #'
 #' # Formula
