+++ R/galamm.R
@@ -182,9 +182,9 @@
 #'   data = mresp,
 #'   family = families,
 #'   family_mapping = family_mapping,
-#'   factor = list("level"),
+#'   factor = "level",
 #'   load.var = "itemgroup",
-#'   lambda = list(loading_matrix)
+#'   lambda = loading_matrix
 #' )
 #'
 #' # Summary information
@@ -224,8 +224,8 @@
 #'     (0 + loading | id),
 #'   data = dat,
 #'   load.var = "item",
-#'   lambda = list(loading_matrix),
-#'   factor = list("loading")
+#'   lambda = loading_matrix,
+#'   factor = "loading"
 #' )
 #'
 #' # We can plot the estimated smooth term
@@ -234,10 +234,10 @@
 #'
 #' # Interaction between observed and latent covariates ------------------------
 #' # Define the loading matrix
-#' lambda <- list(matrix(c(1, NA, NA), ncol = 1))
+#' lambda <- matrix(c(1, NA, NA), ncol = 1)
 #'
 #' # Define the regression functions, one for each row in the loading matrix
-#' factor_interactions <- list(list(~1, ~1, ~x))
+#' factor_interactions <- list(~1, ~1, ~x)
 #'
 #' # Fit the model
 #' mod <- galamm(
@@ -245,7 +245,7 @@
 #'   data = latent_covariates,
 #'   load.var = "type",
 #'   lambda = lambda,
-#'   factor = list("loading"),
+#'   factor = "loading",
 #'   factor_interactions = factor_interactions
 #' )
 #'
