+++ man/factor_loadings.galamm.Rd
@@ -37,7 +37,7 @@ loading_matrix <- matrix(c(1, NA, NA), ncol = 1)
 # Estimate model
 mod <- galamm(y ~ item + (0 + ability | sid) + (0 + ability | school),
   data = IRTsub, family = binomial, load.var = "item",
-  factor = list("ability"), lambda = list(loading_matrix)
+  factor = "ability", lambda = loading_matrix
 )
 
 # Show estimated factor loadings, with standard errors

[…]
