### 2023-10-18 #614 @noamross
Thank you @osorensen for this pre-submission inquiry! This package is well within our statistical scope and we'd be happy to review a full submission. While we recommend *autotest*, documentation of your standards compliance with *srr* and submission are not dependent on a clean *autotest* run.

@mpadge, can you provide some interpretation of this autotest result?
