+++ R/RcppExports.R
@@ -1,6 +1,194 @@
 # Generated by using Rcpp::compileAttributes() -> do not edit by hand
 # Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
 
+#' @title Set up parameter and model family
+#'
+#' @description
+#' Templated wrapper function which sets up the necessary parameters to
+#' evaluate the marginal likelihood. The template type \code{T} will typically
+#' be one of \code{double}, \code{autodiff::dual1st}, and
+#' \code{autodiff::dual2nd}.
+#'
+#' @srrstats {G1.4a} Internal function documented.
+#'
+#' @param y Double precision vector of response values.
+#' @param trials Double precision vector with number of trials. When trials
+#' are not applicable, e.g., with Gaussian or Poisson responses, this should
+#' be a vector of ones.
+#' @param X Fixed effect model matrix.
+#' @param Zt Transpose of random effect model matrix.
+#' @param Lambdat Lower Cholesky factor of random effect covariance matrix.
+#' @param beta Double precision vector of fixed effects.
+#' @param theta Double precision vector with the unique elements of
+#'   \code{Lambdat}.
+#' @param theta_mapping Integer vector mapping elements of \code{theta} to the
+#'   positions in \code{Lambdat}.
+#' @param u_init Double precision vector with initial values of random
+#'   effects. These random effects should be standardized.
+#' @param lambda Double precision vector of factor loadings.
+#' @param lambda_mapping_X Integer vector mapping elements of
+#'   \code{lambda} to elements of \code{X}, in row-major order.
+#' @param lambda_mapping_Zt List of integer vectors mapping elements of
+#'   \code{lambda} to non-zero elements of \code{Zt} assuming compressed
+#'   sparse column format is used. If \code{lambda_mapping_Zt_covs} is of
+#'   length zero, then each list element in \code{lambda_mapping_Zt} should be
+#'   of length one, and it will then be multiplied by the corresponding element
+#'   of \code{Zt}.
+#' @param lambda_mapping_Zt_covs List of double precision vector. Must either
+#'   be of length zero, or the same length as \code{lambda_mapping_Zt_covs}.
+#'   Each list element contains potential covariates that the elements of
+#'   \code{lambda_mapping_Zt} should be multiplied with. If the list is of
+#'   length 0, all elements of \code{lambda_mapping_Zt} are implicitly
+#'   multiplied by 1.
+#' @param weights Double precision vector of weights, used in heteroscedastic
+#'   models.
+#' @param weights_mapping Integer vector mapping the elements of \code{weights}
+#'   to the rows of \code{X}.
+#' @param family Vector of strings defining the family or families. Each
+#'   vector element must currently be one of \code{"gaussian"},
+#'   \code{"binomial"}, or \code{"poisson"}.
+#' @param family_mapping Integer vector mapping elements of \code{family} to
+#'   the rows of \code{X}.
+#' @param k Double precision vector with pre-computed constant term in the
+#'   log-likelihood for each element in \code{family}.
+#' @param maxit_conditional_modes Integer specifying the maximum number of
+#'   iteration in penalized iteratively reweighted least squares algorithm
+#'   used to find the conditional modes of the random effects.
+#' @param lossvalue_tol Double precision scalar specifying the absolute
+#'   convergence criterion for the penalized iteratively reweighted least
+#'   squares algorithm used to find the conditional modes of the random
+#'   effects.
+#' @param reduced_hessian Boolean specifying whether the Hessian matrix of
+#'   second derivatives should be computed only with respect to \code{beta}
[…]
