+++ src/update_funs.h
@@ -23,25 +23,20 @@ void update_Lambdat(SpMdual<T>& Lambdat, Vdual<T> theta,
 
 template <typename T>
 void update_X(Mdual<T>& X, const Vdual<T>& lambda,
-              const std::vector<std::vector<int>>& lambda_mapping_X){
+              const Eigen::VectorXi& lambda_mapping_X){
   if(lambda_mapping_X.size() == 0) return;
   if(lambda_mapping_X.size() != X.size()) Rcpp::stop("Mismatch in lambda_mapping_X size.");
 
   for(size_t i{}; i < lambda_mapping_X.size(); i++){
-    std::vector<int> newinds = lambda_mapping_X[i];
+    int newind = lambda_mapping_X[i];
     T loading{0};
     bool update{false};
-    int j{0};
-
-    for(int newind : newinds){
-      if(newind == -2){
-        loading = 0;
-        update = true;
-      } else if(newind >= 0){
-        loading += lambda(newind);
-        update = true;
-      }
-      j++;
+    if(newind == -2) {
+      loading = 0;
+      update = true;
+    } else if(newind >= 0) {
+      loading += lambda(newind);
+      update = true;
     }
 
     if(update){
