+++ R/mgcv-functions.R
@@ -283,6 +283,8 @@ variable.summary <- function(pf, dl, n) {
     if (v.name[i] %in% p.name) para <- TRUE else para <- FALSE
 
     x <- dl[[v.name[i]]]
+    #' @srrstats {G2.4d} *explicit conversion to factor via `as.factor()`*
+    #' @noRd
     if (is.character(x)) x <- as.factor(x)
     if (is.factor(x)) {
       x <- x[!is.na(x)]

[…]
