+++ src/update_funs.h
@@ -3,6 +3,23 @@
 
 #include "model.h"
 
+//' Update Cholesky factor
+//'
+//' Updates the Cholesky factor of the covariance matrix of the random effects
+//' based on the current values of \code{theta}. Template is typically one of
+//' \code{double}, \code{autodiff:dual1st}, or \code{autodiff::dual2nd}.
+//'
+//' @srrstats {G1.4a} Internal function documented.
+//'
+//' @param Lambda Lower Cholesky factor, an object of class
+//'   \code{Eigen::SparseMatrix<T>}.
+//' @param theta Vector of unique elements of the Cholesky factor, an object of
+//'   class \code{Eigen::Matrix<T, Eigen::Dynamic, 1>}.
+//' @param theta_mapping Integer vector mapping elements of \code{theta} to the
+//'   positions in \code{Lambdat}.
+//' @return No return value. \code{Lambdat} is modified in-place.
+//'
+//' @noRd
 template <typename T>
 void update_Lambdat(SpMdual<T>& Lambdat, Vdual<T> theta,
                     const std::vector<int>& theta_mapping
@@ -21,6 +38,23 @@ void update_Lambdat(SpMdual<T>& Lambdat, Vdual<T> theta,
   }
 };
 
+//' Update fixed effect matrix
+//'
+//' Updates the fixed effect design matrix \code{X} based on the current values
+//' of the factor loadings in \code{lambda}. Template is typically one of
+//' \code{double}, \code{autodiff:dual1st}, or \code{autodiff::dual2nd}.
+//'
+//' @srrstats {G1.4a} Internal function documented.
+//'
+//' @param X Design matrix of class
+//'   \code{Eigen::Matrix<T, Eigen::Dynamic, Eigen::Dynamic>}.
+//' @param lambda Vector of factor loadings, of class
+//'   \code{Eigen::Matrix<T, Eigen::Dynamic, 1>}.
+//' @param lambda_mapping_X Integer vector mapping elements of
+//'   \code{lambda} to elements of \code{X}, in row-major order.
+//' @return No return value. \code{X} is modified in-place.
+//'
+//' @noRd
 template <typename T>
 void update_X(Mdual<T>& X, const Vdual<T>& lambda,
               const Eigen::VectorXi& lambda_mapping_X){
@@ -47,6 +81,33 @@ void update_X(Mdual<T>& X, const Vdual<T>& lambda,
   }
 };
 
+//' Update random effect matrix
+//'
+//' Updates the random effect design matrix \code{X} based on the current values
+//' of the factor loadings in \code{lambda}. Template is typically one of
+//' \code{double}, \code{autodiff:dual1st}, or \code{autodiff::dual2nd}.
+//'
+//' @srrstats {G1.4a} Internal function documented.
+//'
+//' @param Zt Transpose of design matrix for random effects, of class
+//'   \code{Eigen::SparseMatrix<T>}.
+//' @param lambda Vector of factor loadings, of class
+//'   \code{Eigen::Matrix<T, Eigen::Dynamic, 1>}.
+//' @param lambda_mapping_Zt Vector of integer vectors mapping elements of
+//'   \code{lambda} to non-zero elements of \code{Zt} assuming compressed
+//'   sparse column format is used. If \code{lambda_mapping_Zt_covs} is of
+//'   length zero, then each outer element in \code{lambda_mapping_Zt} should be
+//'   of length one, and it will then be multiplied by the corresponding element
+//'   of \code{Zt}.
+//' @param lambda_mapping_Zt_covs Vectir of double precision vector. Must either
+//'   be of length zero, or the same length as \code{lambda_mapping_Zt_covs}.
[…]
