### 2023-10-19 eee95c6
moved test to where it belongs
